Move sprites out of slides directory
