+++ src/components/bootstrap-modal/bootstrap-modal.ts
@@ -96,8 +96,8 @@ export class VerificationBootstrapComponent extends AbstractComponent(LitElement
       // new PagingSlide(),
       // new ShortcutsSlide(this.decisionShortcuts, this),
 
-      new SelectionSlide(),
       new ShortcutsSlide(this.decisionShortcuts, this),
+      new SelectionSlide(),
       new PagingSlide(),
       new DecisionsSlide(this.hasVerificationTask, this.hasClassificationTask),
     ];
