+++ src/components/bootstrap-modal/slides/selection/animations.css
@@ -1,4 +1,6 @@
 .selection-slide {
+  --animation-duration: 5s;
+
   @keyframes selection-cursor-movement {
     from {
       transform: translate3d(-70px, -60px, 0);
@@ -26,8 +28,45 @@
       transform: none;
     }
   }
+
+  @keyframes highlight-box-animation {
+    to {
+      width: 100%;
+    }
+  }
+
+  @keyframes selected-animation {
+    from {
+      --oe-panel-color: var(--oe-selected-color);
+    }
+  }
+
+  @keyframes selected-animation-highlight {
+    from {
+      --oe-panel-color: var(--oe-selected-color);
+    }
+  }
 }
 
 .selection-cursor {
-  animation: selection-cursor-movement 5s infinite;
+  animation: selection-cursor-movement var(--animation-duration) infinite;
+}
+
+.selection-box {
+  animation: highlight-box-animation var(--animation-duration) infinite;
+}
+
+.tile-1 {
+  animation: selected-animation var(--animation-duration) infinite;
+  animation-delay: 0.5s;
+}
+
+.tile-2 {
+  animation: selected-animation var(--animation-duration) infinite;
+  animation-delay: 1.5s;
+}
+
+.tile-5 {
+  animation: selected-animation var(--animation-duration) infinite;
+  animation-delay: 2.5s;
 }
