+++ src/components/bootstrap-modal/slides/selection/selection.ts
@@ -1,10 +1,10 @@
 import { importSprites } from "../../../../helpers/svgs/imports";
 import { AbstractSlide } from "../abstractSlide";
 import { svg } from "lit";
-import { verificationGridSprite } from "../sprites/verification-grid.sprite";
-import gridTileSprite from "../sprites/grid-tile.svg?raw";
-import cursorSprite from "../sprites/cursor.svg?raw";
-import selectionBoxSprite from "../sprites/selection-box.svg?raw";
+import { verificationGridSprite } from "../../sprites/verification-grid.sprite";
+import gridTileSprite from "../../sprites/grid-tile.svg?raw";
+import cursorSprite from "../../sprites/cursor.svg?raw";
+import selectionBoxSprite from "../../sprites/selection-box.svg?raw";
 
 export class SelectionSlide extends AbstractSlide {
   public constructor() {
@@ -20,6 +20,7 @@ export class SelectionSlide extends AbstractSlide {
 
         <use href="#cursor" class="selection-cursor" x="225" y="100" />
         <use
+          class="selection-box"
           href="#selection-box"
           x="25"
           y="20"
