+++ src/components/bootstrap-modal/slides/shortcuts/animations.css
@@ -1,8 +1,5 @@
 .shortcut-slide {
   @keyframes selection-animation {
-    from {
-    }
-
     to {
       --oe-panel-color: var(--oe-selected-color);
     }
