+++ src/components/bootstrap-modal/sprites/types.ts
